Done Styling of campaign cards, aligned with the other elements on the page
