+++ app/assets/stylesheets/pages/_home.scss
@@ -1,3 +1,10 @@
+// Container for centering content
+
+.container {
+  display: flex;
+  justify-content: space-around;
+}
+
 // Banner
 
 .banner {
@@ -49,9 +56,6 @@
   .picture{
     width: 650px;
     height:99.9%;
-    // position: absolute;
-    // left: -0.5px;
-    // top: -0.5px;
   }
 }
 
@@ -75,22 +79,19 @@ text-decoration: none;
 
 }
 
-// Explanation business
-
-.business-explanation {
-  margin: 0 auto;
-}
+// Explanation of aim
 
 .how-wrap {
   width: 1300px;
   margin: 150px 150px 0 150px;
-  border:solid black .5px;
+  border: solid black .5px;
 }
 
 .how-wrap1 {
   width: 1300px;
   margin: -0.9px 150px;
-  border:solid black .5px;
+  border-left: solid black .5px;
+  border-right: solid black .5px;
 }
 
 .how-wrap2 {
@@ -124,7 +125,7 @@ margin-right: 24px;
 margin-top: 30px;
 }
 
-.link-t, .link-t1 {
+.link-t {
 padding-top:4px;
 font-size:25px;
 margin-right:36px;
@@ -137,48 +138,49 @@ background-color: #e7daf5;
 padding:10px 0;
 }
 
+.para {
+  background-color: #e7daf5;
+  padding:.1px 0;
+  border-right: solid black .5px;
+  border-left: solid black .5px;
+}
+
 // Selection campaigns
 
-.card-flex {
-display:flex;
-}
-
-.card-campaign {
-background-color: white;
-width: 333px;
-margin:50px;
-font-size: 19px;
-line-height: 124.9%;
-letter-spacing: 0.03em;
-text-indent: 1px;
-}
-
-.para {
-background-color: #e7daf5;
-padding:.1px 0;
-}
-
-
-.mini-banner {
-text-align:center;
-padding: 55px 100%;
-background-color: F9632E;
-}
-
-
+// Purple Banner
 .latest-campaign {
-display:flex;
-align-items: center;
-justify-content: center;
-width:100%;
-height:95px;
-background-color: #5c4aff;
-margin-bottom: 100px;
-color:white;
+  display:flex;
+  align-items: center;
+  justify-content: center;
+  width:100%;
+  height:95px;
+  background-color: #5c4aff;
+  margin-bottom: 100px;
+  color:white;
 }
 
 .latest-c {
-font-size: 23px;
+  font-size: 23px;
+}
+
+// Overview three campaigns
+
+.card-flex {
+  display:flex;
+  justify-items: space-around;
+  width: 1300px;
+}
+
+.card-campaign {
+  background-color: white;
+  width: 333px;
+  margin:50px;
+  padding-right: 15px;
+  padding-left: 15px;
+  font-size: 19px;
+  line-height: 124.9%;
+  letter-spacing: 0.03em;
+  text-indent: 1px;
 }
 
 .card-campaign-infos {
@@ -205,14 +207,14 @@ font-size: 23px;
   font-weight: bold;
 }
 
-
+// Link to all campaigns
 .more-campaign-button {
   display: flex;
   align-items: center;
   justify-content: space-around;
-  margin-bottom: 300px;
-
+  padding-top: 100px;
 }
+
 .more-button {
   border: .5px black solid;
   font-size: 17px;
@@ -223,4 +225,3 @@ font-size: 23px;
   padding: 13px 27px;
   color: black;
 }
-
