+++ app/javascript/packs/application.js
@@ -2,5 +2,7 @@ import "bootstrap";
 import 'mapbox-gl/dist/mapbox-gl.css'; // <-- you need to uncomment the stylesheet_pack_tag in the layout!
 
 import { initMapbox } from '../plugins/init_mapbox';
-
 initMapbox();
+
+import { toggleExplanations } from './home.js';
+toggleExplanations();
